refactor: Merge + api-extract.json + api-deprecations.json
